requirements.txt
Django==4.2.11
mysqlclient==2.2.4
djangorestframework==3.15.1
django-debug-toolbar==4.3.0